+++ app/user/repository.go
@@ -5,8 +5,8 @@ import (
 )
 
 type Repository interface {
-	Create(user *entities.User) (int64, error)
-	Update(condition *entities.User, user *entities.User) (int64, error)
+	Create(user *entities.User) (uint, error)
+	Update(condition *entities.User, user *entities.User) (uint, error)
 	List(skip int, limit int) ([]*entities.User, error)
 	GetByPhoneNumber(id int64) (entities.User, error)
 }
